+++ qmtl/gateway/__init__.py
@@ -2,7 +2,8 @@ from .dagmanager_client import DagManagerClient
 from .redis_queue import RedisTaskQueue
 from .redis_client import InMemoryRedis
 from .worker import StrategyWorker
-from .api import create_app, StrategySubmit, StrategyAck, StatusResponse
+from .api import create_app
+from .models import StrategySubmit, StrategyAck, StatusResponse
 from .degradation import DegradationManager, DegradationLevel
 from .database import Database
 from .fsm import StrategyFSM
