+++ tests/gateway/test_api.py
@@ -1,7 +1,8 @@
 import pytest
 import httpx
 
-from qmtl.gateway.api import create_app, Database, StrategySubmit
+from qmtl.gateway.api import create_app, Database
+from qmtl.gateway.models import StrategySubmit
 from qmtl.common import crc32_of_list
