+++ tests/gateway/test_nodeid.py
@@ -5,7 +5,8 @@ import pytest
 from fastapi.testclient import TestClient
 
 from qmtl.common import compute_node_id, crc32_of_list
-from qmtl.gateway.api import create_app, Database, StrategySubmit
+from qmtl.gateway.api import create_app, Database
+from qmtl.gateway.models import StrategySubmit
 
 
 class FakeDB(Database):
